requirements.txt
pandas
yfinance
python-dotenv
sqlalchemy
pyodbc